separate script/InPageElement.ts into inPageElement/*
