+++ src/inPageElement/getVideo.ts
@@ -1,0 +1,9 @@
+import waitAppearance from "./function/waitAppearance";
+
+const getVideo = () =>
+  waitAppearance(() =>
+    document
+      .querySelector<HTMLVideoElement>('.video-stream')
+  );
+
+export default getVideo;
